chore: added old version of backend for WPFs
In the SIMS legacy code, the Backend was a class library that the WPF frontends reference directly, which impedes current development and refactoring on the Backend. The LegacyBackend was added so the WPF frontends can be detached from the current Backend, and it will be deleted when HTTP communication is fully established between the WPF frontends and the Backend.