refactor[d73]: registration refactored to work through services
I am aware that the code in this commit doesn't adhere to our clean code principles, however, currently the priority is separating the PatientWebApp from Backend. A branch has been made to refactor the HTTP communication in PatientWebApp to be cleaner, and the registration endpoint needs to be reworked again anyway since saving the images in the current way doesn't properly work on Heroku.